Annotate mutations target motif

Checks for the presence of mutations at a given motif

annotate_ssm_motif_context(
  maf,
  motif = "WRCY",
  index = 3,
  genome_build,
  fastaPath
)

Arguments

maf

MAF data frame (required columns: Reference_Allele, Chromosome, Start_Position, End_Position)

motif

The motif sequence (default is WRCY)

index

Position of the mutated allele in the motif

genome_build

The genome build for the variants you are working with (default is grch37)

fastaPath

Can be a path to a FASTA file

Value

A data frame with two extra columns (seq and motif).

Details

In positions that reference allele has been mutated, it will capture (motif length - 1) before and (motif length + 1) alleles after the mutated position. Then, it looks for the presence of motif in the captured sequence and check if the mutation has occurred in the indexed position, it will return SITE and if the the motif is present, but the mutation is not in the indexed position, it will return MOTIF. In other cases, it will return FALSE. NA will be shown if the mutation is an indel mutation.

Examples

my_maf <- get_coding_ssm() %>% head()
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)

annotated = annotate_ssm_motif_context(maf = my_maf, motif = "WRCY", index = 3)

print(annotated)
#> genomic_data Object
#> Genome Build: grch37 
#> Showing first 10 rows:
#>     Hugo_Symbol Entrez_Gene_Id Center NCBI_Build Chromosome Start_Position
#> 1    AL627309.1              0      .     GRCh37          1         138626
#> 2    AL627309.1              0      .     GRCh37          1         138972
#> 3 RP11-206L10.9              0      .     GRCh37          1         730845
#> 4        FAM87B              0      .     GRCh37          1         753589
#> 5        SAMD11              0      .     GRCh37          1         871158
#> 6        SAMD11              0      .     GRCh37          1         871192
#>   End_Position Strand Variant_Classification Variant_Type Reference_Allele
#> 1       138626      +                 Silent          SNP                T
#> 2       138973      +        Frame_Shift_Ins          INS                -
#> 3       730845      +          Splice_Region          SNP                G
#> 4       753589      +          Splice_Region          SNP                A
#> 5       871158      +                 Silent          SNP                C
#> 6       871192      +      Missense_Mutation          SNP                C
#>   Tumor_Seq_Allele1 Tumor_Seq_Allele2    dbSNP_RS dbSNP_Val_Status
#> 1                 T                 C       novel               NA
#> 2                 -                GC rs760414759               NA
#> 3                 G                 A rs368628623               NA
#> 4                 A                 C       novel               NA
#> 5                 C                 T rs770660212               NA
#> 6                 C                 T rs548621473               NA
#>        Tumor_Sample_Barcode Matched_Norm_Sample_Barcode Match_Norm_Seq_Allele1
#> 1                 09-25670T                   09-25670N                      T
#> 2 BLGSP-71-30-00673-01A-01E   BLGSP-71-06-00286-99A-01D                      -
#> 3                  SP116668                    SP116667                      G
#> 4                      EB-3   BLGSP-71-06-00286-99A-01D                      A
#> 5                 05-33344T                   14-11247N                      C
#> 6 BLGSP-71-30-00662-01A-01E   BLGSP-71-06-00286-99A-01D                      C
#>   Match_Norm_Seq_Allele2 Tumor_Validation_Allele1 Tumor_Validation_Allele2
#> 1                      T                       NA                       NA
#> 2                      -                       NA                       NA
#> 3                      G                       NA                       NA
#> 4                      A                       NA                       NA
#> 5                      C                       NA                       NA
#> 6                      C                       NA                       NA
#>   Match_Norm_Validation_Allele1 Match_Norm_Validation_Allele2
#> 1                            NA                            NA
#> 2                            NA                            NA
#> 3                            NA                            NA
#> 4                            NA                            NA
#> 5                            NA                            NA
#> 6                            NA                            NA
#>   Verification_Status Validation_Status Mutation_Status Sequencing_Phase
#> 1                  NA                NA              NA               NA
#> 2                  NA                NA              NA               NA
#> 3                  NA                NA              NA               NA
#> 4                  NA                NA              NA               NA
#> 5                  NA                NA              NA               NA
#> 6                  NA                NA              NA               NA
#>   Sequence_Source Validation_Method Score BAM_File Sequencer Tumor_Sample_UUID
#> 1              NA                NA    NA       NA        NA                NA
#> 2              NA                NA    NA       NA        NA                NA
#> 3              NA                NA    NA       NA        NA                NA
#> 4              NA                NA    NA       NA        NA                NA
#> 5              NA                NA    NA       NA        NA                NA
#> 6              NA                NA    NA       NA        NA                NA
#>   Matched_Norm_Sample_UUID          HGVSc             HGVSp HGVSp_Short
#> 1                       NA       c.684A>G               p.=     p.S228=
#> 2                       NA c.337_338insGC p.Pro113ArgfsTer3 p.P113Rfs*3
#> 3                       NA     n.259-5G>A              <NA>        <NA>
#> 4                       NA     n.832+7N>C              <NA>        <NA>
#> 5                       NA       c.312C>T               p.=     p.D104=
#> 6                       NA       c.346N>T       p.Arg116Cys     p.R116C
#>     Transcript_ID Exon_Number t_depth t_ref_count t_alt_count n_depth
#> 1 ENST00000423372         1/2      69          48          21      35
#> 2 ENST00000423372         1/2      52          31          21      36
#> 3 ENST00000586288        <NA>      31          21          10      23
#> 4 ENST00000326734        <NA>      97          47          50      35
#> 5 ENST00000342066        5/14      93          42          51      16
#> 6 ENST00000342066        5/14      20          16           4      51
#>   n_ref_count n_alt_count maf_seq_type     seq  WRCY
#> 1          35           0       genome CCATGAG FALSE
#> 2          36           0       genome      NA    NA
#> 3          23           0       genome TTCGTTA FALSE
#> 4          35           0       genome AGAAATT FALSE
#> 5          16           0       genome GGACGGC FALSE
#> 6          51           0       genome CACCGCA FALSE