ASHM Multi-panel Rainbow Plot

Generates a colourful multi-panel overview of hypermutation in regions of interest across many samples.

ashm_multi_rainbow_plot(
  regions_bed,
  regions_to_display,
  exclude_classifications,
  metadata,
  this_seq_type,
  custom_colours,
  classification_column = "lymphgen",
  maf_data,
  projection = "grch37",
  verbose = TRUE
)

Arguments

regions_bed

Bed file with chromosome coordinates, should contain columns chr, start, end, name (with these exact names). Not required if selecting from many common regions; bonus regions also exist in grch37.

regions_to_display

Optional vector of names from default regions_bed to use.

exclude_classifications

Optional argument for excluding specific classifications from a metadata file.

metadata

A metadata file already subsetted and arranged on the order you want the samples vertically displayed.

this_seq_type

the seqtype you want results back for if maf_data is not provided.

custom_colours

Provide named vector (or named list of vectors) containing custom annotation colours if you do not want to use standardized pallette.

classification_column

Optional. Override default column for assigning the labels used for colouring in the figure.

maf_data

An already loaded maf. If not provided, this function will call get_ssm_by_region, using the regions supplied into regions_bed. Ensure your maf matches the genome projection.

projection

Provide genome build; default is grch37. Bonus regions are only available in grch37.

verbose

Set to FALSE to prevent printing the full regions bed file to the console. Default is TRUE.

Value

Nothing

Details

The input for this function is a bed-file with the following columns; chr, start, end, name. Note that for this function to work, the column names must be exactly this. The user also needs to specify a vector of names (regions_to_display) to further control what regions are to be displayed on the returned plot. It is also possible to exclude specific classifications from the metadata file. This is achieved with exclude_classifications. In addition the user can also use the metadata parameter to use an already subset and arranged metadata table. This function will call get_ssm_by_region if maf_data is not called. For more info, refer to the parameter descriptions of this function.

Examples

library(GAMBLR.data)

#get lymphgen colours
lymphgen_colours = GAMBLR.helpers::get_gambl_colours(classification = "lymphgen")

#build plot
if (FALSE) { # \dontrun{
ashm_multi_rainbow_plot(regions_to_display = c("BCL2-TSS",
                                               "MYC-TSS",
                                               "SGK1-TSS",
                                               "IGL"),
                        custom_colours = lymphgen_colours,
                        this_seq_type = "genome")
} # }