Create cBioPortal Study.

Wrapper function for creating a import-ready cBioPortal study.

Usage

cbio_create(
  these_samples_metadata = NULL,
  maf_data = NULL,
  this_seq_type = "genome",
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_genome",
  description = "GAMBL data from genome",
  gambl_maf = "maf_slms3_hg19",
  gambl_icgc_maf = "maf_slms3_hg19_icgc",
  cancer_type = "mixed",
  overwrite = TRUE,
  check_study = TRUE,
  out_dir
)

Arguments

these_samples_metadata

Metadata for the samples to be included in the study.

maf_data

Data frame with maf data for the samples in the study.

this_seq_type

The seq type you want to generate a study for. Default is "genome".

short_name

A concise name for your portal project. Default is "GAMBL".

human_friendly_name

A slightly more verbose name for your project. Default is "GAMBL data".

project_name

Unique ID for your project. Default is "gambl_all".

description

A verbose description of your data set. This is what the study will be named when accessing it through cBioPortal. Default is "GAMBL data from genome".

gambl_maf

MAF origin.

gambl_icgc_maf

ICGC MAF origin.

cancer_type

Cancer types included in study, default is "mixed".

overwrite

Flag to specify that files should be overwritten if they exist. Default is TRUE.

check_study

Boolean parameter that controls if the generated study should be checked for sample IDs in case lists, that are not described in the clinical file. Default is TRUE.

out_dir

The full path to the base directory where the files are being created.

Value

Nothing. Rather, this function generates all files necessary for successfully importing a study into an active cBioPortal instance.

Details

This function internally calls cbio_setup_study, cbio_setup_fusions, cbio_finalize_study and cbio_study_check to generate all necessary files for importing a study into cBioPortal. This function was developed to streamline this step and at the same time ensure that the study information and selected data type is consistent throughout the individual steps of generating a study. In addition, the user can also control if the generated study should be checked for sample IDs in case lists that are not described in the clinical file. This potentially will prevent an annoying error that prevents the study to be imported into the active cBioPortal instance, default is TRUE. Fusions are also handled based on the selected seq type (this_seq_type).

Examples

if (FALSE) { # \dontrun{
#generate cBioPortal study for all GAMBL genome samples:
cbio_create()

#generate a cBioPortal study for all GAMVL capture samples:
cbio_create(this_seq_type = "capture", description = "GAMBL data from exomes")
} # }