Setup Fusions (cBioPortal).

Annotate SVs and create the input for fusions to be displayed in cBioPortal instance.

Usage

cbio_setup_fusions(
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_genome",
  description = "GAMBL data from genome",
  gambl_maf = "maf_slms3_hg19",
  gambl_icgc_maf = "maf_slms3_hg19_icgc",
  out_dir
)

Arguments

short_name

A concise name for your portal project.

human_friendly_name

A slightly more verbose name for your project.

project_name

Unique ID for your project.

description

A verbose description of your data set.

gambl_maf

maf origin.

gambl_icgc_maf

ICGC maf origin.

out_dir

The full path to the base directory where the files are being created.

Value

A vector of sample_id for the patients that have been included.

Details

This function calls get_combined_sv and runs annotation on the returned data frame. Should be run as the next step after running cbio_setup_study. Note that the parameters called with this function has to match the parameter arguments of cbio_setup_study, i.e if short_name is for cbio_setup_study is "GAMBL", then the short_name in cbio_setup_fusions also has to be "GAMBL", etc.

Examples

if (FALSE) { # \dontrun{
fusion_ids = cbio_setup_fusions(out_dir = "GAMBLR/cBioPortal/instance01/")
} # }