Setup Study (cBioPortal).

Initialize a new cBioPortal instance or update existing portal data set, can also be used to retrieve sample ids included in study.

Usage

cbio_setup_study(
  these_samples_metadata = NULL,
  maf_data = NULL,
  seq_type_filter = "genome",
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_genome",
  description = "GAMBL data from genome",
  overwrite = TRUE,
  out_dir
)

Arguments

these_samples_metadata

Metadata for the samples to be included in the study.

maf_data

Data frame with maf data for the samples in the study.

seq_type_filter

the seq type you are setting up a study for, default is "genome".

short_name

A concise name for your portal project.

human_friendly_name

A slightly more verbose name for your project.

project_name

Unique ID for your project.

description

A verbose description of your data set.

overwrite

Flag to specify that files should be overwritten if they exist. Default is TRUE.

out_dir

The full path to the base directory where the files are being created.

Value

A vector of sample_id for the patients that have been included.

Details

This function internally calls get_coding_ssm to retrieve coding mutations to be included in the study (if overwrite = TRUE). Optionally, provide the data frame with somatic mutations with the maf_data argument. In addition, this function also creates and sets up the proper folder hierarchy and writes the files necessary to import a new cBioPortal study. Before a study is ready to be imported to cBioPortal, the user also needs to run cbio_setup_fusions and cbio_finalize_study. Optionally the user can also run cbio_study_check to ensure all samples described by the "clinical" file are included in the study. Also, note that the parameters chosen for this function have to match the same parameters called for any subsequent study function calls.

Examples

#Setup study and save included ids as a vector of characters:
if (FALSE) { # \dontrun{
ids = cbio_setup_study(out_dir = "GAMBLR/cBioPortal/instance01/")
} # }