Fetch Output Files.

Get output files from a set of conditions.

Usage

fetch_output_files(
  tool,
  unix_group,
  base_path,
  results_dir = "99-outputs",
  this_seq_type = "genome",
  build = "hg38",
  search_pattern = "cellularity_ploidy.txt"
)

Arguments

tool

Name of tool.

unix_group

The unix group of the sample set.

base_path

Either the full or relative path to where all the results directories are for the tool e.g. "gambl/sequenza_current".

results_dir

Directory with results.

this_seq_type

Either genome or capture.

build

Default is hg38.

search_pattern

File-extensions search pattern.

Value

A data frame with one row per file and sample IDs parsed from the file name along with other GAMBL wildcards.

Details

This function lets the user specify multiple conditions for returning result subsets. First, specify the name of the tool with tool, then set the seq type (this_seq_type) to either genome or capture, together with the genome build (genome_build). A data frame will be returned with one row per file and sample IDs together with GAMBL wildcards.

Examples

if (FALSE) { # \dontrun{
ex_outs = fetch_output_files(tool = "manta",
                             base_path = "gambl/sequenza_current",
                             this_seq_type = "capture",
                             build = "hg38")
} # }