Get GAMBL Outcomes.

Get the patient-centric clinical metadata.

Usage

get_gambl_outcomes(
  patient_ids,
  time_unit = "year",
  censor_cbioportal = FALSE,
  complete_missing = FALSE,
  from_flatfile = TRUE
)

Arguments

patient_ids

Vector of patient IDs.

time_unit

Return follow-up times in one of three time units: year, month or day. Default is "year".

censor_cbioportal

Optionally request the censoring to be encoded in the specific style required by cBioPortal. Default is FALSE.

complete_missing

Optionally fill in any gaps to ensure we have values for every patient (censor at 0 if missing). Default is FALSE.

from_flatfile

Optionally set to FALSE to use the database to get the survival data. Default is TRUE.

Value

Data frame with one row for each patient_id.

Details

INTERNAL FUNCTION called by get_gambl_metadata, not meant for out-of-package usage.

Examples

if (FALSE) { # \dontrun{
  outcome_df = get_gambl_outcomes()
  metadata_with_outcomes = dplyr::left_join(
                            get_gambl_metadata(),
                            outcome_df)
} # }