Get Manta SV By Samples.

Load the manta output for a set of samples.

Usage

get_manta_sv_by_samples(
  these_samples_metadata,
  min_vaf = 0.1,
  min_score = 40,
  pass_filters = TRUE,
  projection = "grch37",
  verbose = TRUE
)

Arguments

these_samples_metadata

The only required parameter is a metadata table (data frame) that must contain a row for each sample you want the data from. The additional columns the data frame needs to contain, besides sample_id, are: unix_group, genome_build, seq_type, pairing_status.

min_vaf

The minimum tumour VAF for a SV to be returned. Default value is 0.1.

min_score

The lowest Manta somatic score for a SV to be returned. Default value is 40.

pass_filters

If set to TRUE, only return SVs that are annotated with PASS in the FILTER column. Set to FALSE to keep all variants, regardless if they PASS the filters. Default is TRUE.

projection

The projection of returned calls. Default is grch37.

verbose

Set to FALSE to prevent the path of the requested bedpe file to be printed.

Value

A data frame containing the Manta outputs from all sample_id in these_samples_metadata in a bedpe-like format with additional columns extracted from the VCF column.

Details

This is a convenience wrapper function for get_manta_sv_by_sample (and called by get_manta_sv). Is this function not what you are looking for? Try one of the following, similar, functions; get_combined_sv, get_manta_sv, get_manta_sv_by_sample

Examples

if (FALSE) { # \dontrun{
all_sv = get_manta_sv()
metadata = get_gambl_metadata()
missing_samples = dplyr::anti_join(metadata,
  all_sv,
  by = c("sample_id" = "tumour_sample_id"))

missing_from_merge = get_manta_sv_by_samples(these_samples_metadata = missing_samples, verbose = FALSE)
} # }