Build UCSC browser track hub

Create a directory that contains the track hub files. They are: a hub.txt file (named as <projection>_hub.txt and marked with useOneFile on), and a subdirectory (named as the projection in use) with the custom tracks to be visualized in UCSC browser. The custom track regions contains the regions from where SSMs were retrieved. All other custom tracks contain SSMs from samples separated by the splitColumnName parameter, where each file is named according to the value in splitColumnName that it refers to.

build_browser_hub(
  maf_data,
  regions_bed,
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  these_seq_types = c("genome", "capture"),
  projection = "grch37",
  local_web_host_dir = NULL,
  hub_dir = "my_hub",
  splitColumnName = "pathology",
  hub_name = basename(hub_dir),
  shortLabel = basename(hub_dir),
  longLabel = basename(hub_dir),
  contact_email,
  visibility = "squish",
  bigDataUrl_base = "https://github.com/morinlab/LLMPP/blob/main",
  bedToBigBed_path
)

Arguments

maf_data

A MAF-format data frame with the SSMs to be visualized in the UCSC browser. It can be filtered by regions and samples provided by parameters regions_bed and these_sample_ids/these_samples_metadata, respectively.

regions_bed

A BED-format table with the regions you want to retrieve SSMs from. The columns 1, 2 and 3 must be chromosome names, start positions and end positions, respectively. If not provided, the function defaults to aSHM regions, i.e. one of GAMBLR.data::grch37_ashm_regions or GAMBLR.data::hg38_ashm_regions, which is chosen automatically depending on the provided projection.

these_sample_ids

A vector of sample IDs that you want results for.

these_samples_metadata

A metadata table (with sample IDs in a column) to subset the samples of interest.

these_seq_types

A vector of one or more seq types you want results for. Possible values are "genome" or "capture". If more than one seq type is provided, for each value in splitColumnName, the function creates a separate track for each provided seq type. The default is c("genome", "capture"). See the Details section for more information.

projection

The projection genome build. One of "grch37" (default) or "hg38".

local_web_host_dir

Path to the directory where is your local copy of the web host to be used. For example, if the hub should be hosted on GitHub, local_web_host_dir should be the path to your local copy of the repository directory. Default is NULL.

hub_dir

Path to the directory (inside a web host) where you want to build your track hub. If this directory does not exist, it will be created.

splitColumnName

An single string to indicate which metadata column is used to split the MAF data into custom track files. Default is "pathology".

hub_name

A string with the hub name (without spaces) to fill in the hub field of the hub.txt file. Default is basename(hub_dir).

shortLabel

A string with the short hub label (maximum of 17 characters recommended; spaces are allowed) to fill in the shortLabel field of the hub.txt file. Default is basename(hub_dir).

longLabel

A string with the long hub label (maximum of 80 characters recommended; spaces are allowed) to fill in the longLabel field of the hub.txt file. Default is basename(hub_dir).

contact_email

Required parameter. A string with the contact email to fill in the email field of the hub.txt file.

visibility

A string that controls the track visibility mode. Possible values are "pack", "dense", "full", or "squish" (default).

bigDataUrl_base

A string with the base path of the web location of the tracks' data files (bigDataUrl fields from the hub.txt file). For example, if the hub should be hosted on GitHub, bigDataUrl_base should be something like "https://github.com/morinlab/LLMPP/blob/main" (the default). See how the track's paths from the bigDataUrl fields are set in the Details section.

bedToBigBed_path

Path to your local bedToBigBed UCSC tool. If missing, GAMBLR.helpers::check_config_value is called internally and the bedToBigBed path is obtained from the config.yml file saved in the current working directory.

Value

Nothing.

Details

build_browser_hub will create a custom track file for each combination of these_seq_types and splitColumnName (if mutations can be found). Custom track files are named as <a_seq_type_value>_<a_splitColumnName_value>.bb.

The bigDataUrl field of a track in the hub.txt file is defined in the following way:

file.path(bigDataUrl_base, hub_dir, projection,
          paste0(track_file_names[i], "?raw=true\n"))

where track_file_names[i] is the custom track file name.

Examples

if (FALSE) { # \dontrun{
# create a track hub in LLMPP GitHub repo

library(GAMBLR.data)

local_web_host_dir = "~/repos/LLMPP"
hub_dir = "hubs/ashm_test"

my_meta = get_gambl_metadata() %>% 
  filter(pathology %in% c("BL", "DLBCL", "FL"))

build_browser_hub(
  these_samples_metadata = my_meta,
  these_seq_types = c("genome", "capture"),
  projection = "grch37",
  local_web_host_dir = local_web_host_dir,
  hub_dir = hub_dir,
  splitColumnName = "pathology",
  longLabel = "Public aSHM mutations separated by pathologies",
  contact_email = "rdmorin@sfu.ca"
)
} # }