Get Combined SV.

Retrieve Combined Manta and GRIDSS-derived SVs.

get_combined_sv(
  these_samples_metadata,
  projection = "grch37",
  oncogenes,
  region,
  min_vaf = 0,
  these_sample_ids
)

Arguments

these_samples_metadata

a GAMBL metadata frame containing only the samples you want returned

projection

The projection genome build. Default is "grch37".

oncogenes

A character vector of genes commonly involved in translocations.

region

Optional, region formatted like chrX:1234-5678 (chromosome can be prefixed or not prefixed) instead of specifying chromosome, start and end separately.

min_vaf

The minimum tumour VAF for a SV to be returned. Recommended: 0. (default: 0)

these_sample_ids

DEPRECATED

You may want to include one or more e.g.: c("CCND1", "BCL2","MYC") NOTE: If you are looking for SV affecting oncogenes you are more likely going to want to pass the full output to GAMBLR.utils::annotate_sv

Value

A data frame in a bedpe-like format with additional columns that allow filtering of high-confidence SVs.

Details

The bedpe files this function loads were pre-filtered for a minimum VAF of 0.05, and SVs affecting common translocation regions (BCL2, BCL6, MYC, CCND1) were whitelisted (e.g. no VAF filter applied). Therefore if you wish to post-filter the SVs we recommend doing so as needed. Further, the input bedpe file is annotated with oncogenes and superenhancers from naive and germinal centre B-cells. You can subset to events affecting certain loci using the "oncogenes" argument. Try one of the following, similar, functions; get_manta_sv

Examples

# lazily get every SV in the table with default quality filters
all_sv <- get_combined_sv()
head(all_sv)
#> genomic_data Object
#> Genome Build: grch37 
#> Showing first 10 rows:
#>   CHROM_A  START_A    END_A CHROM_B  START_B    END_B
#> 1       1  3587650  3587656       1  3622021  3622027
#> 2      10 81676978 81676979      10 92752813 92752814
#> 3      10 85593304 85593641       6  2792479  2792814
#> 4      10 96210548 96210907      12 92608729 92609149
#> 5      10 98828963 98828965      10 98862637 98862639
#> 6      11 47663585 47663928      16 53044937 53045280
#>                             manta_name SCORE STRAND_A STRAND_B tumour_sample_id
#> 1         MantaDEL:11066:0:1:0:0:0_bp1    NA        +        -       H054_T_WGS
#> 2        MantaDEL:132721:0:1:0:1:0_bp1    NA        +        -       H054_T_WGS
#> 3           MantaBND:85878:0:2:0:0:0:1    NA        -        +       H054_T_WGS
#> 4          MantaBND:133537:0:1:0:0:0:1    NA        +        +       H054_T_WGS
#> 5 MantaDUP:TANDEM:133659:0:1:0:0:0_bp1    NA        -        +       H054_T_WGS
#> 6          MantaBND:121876:0:5:0:0:0:0    NA        -        +       H054_T_WGS
#>   normal_sample_id VAF_tumour DP       gridss_name
#> 1       H054_N_WGS      0.165 85   gridss0fb_2691o
#> 2       H054_N_WGS      0.197 71 gridss176fb_2753o
#> 3       H054_N_WGS      0.154 39              <NA>
#> 4       H054_N_WGS      0.114 44              <NA>
#> 5       H054_N_WGS      0.122 90  gridss178bf_220o
#> 6       H054_N_WGS      0.122 41              <NA>
#>                      ANNOTATION_A DIST_TO_ANNOTATION_A
#> 1             SE_TP73,SE_TP73-AS3                    0
#> 2                      SE_TSPAN14               535036
#> 3  SE_C10orf99,SE_CERNA2,SE_GHITM               315834
#> 4         SE_AL157834.2,SE_PDLIM1               779450
#> 5 SE_PIK3AP1,SE_TM9SF3,SE_PIK3AP1              -300156
#> 6              SE_FNBP4,SE_NUP160                76324
#>                                                                                        ANNOTATION_B
#> 1                                                                               SE_TP73,SE_TP73-AS3
#> 2                                                                       SE_AL365434.1,SE_AL365434.2
#> 3 SE_MIR4645,SE_SERPINB1,SE_SERPINB9,SE_SERPINB9P1,SE_MIR4645,SE_SERPINB1,SE_SERPINB9,SE_SERPINB9P1
#> 4                             SE_AC025164.1,SE_BTG1,SE_LINC01619,SE_AC025164.1,SE_BTG1,SE_LINC01619
#> 5                                                                   SE_PIK3AP1,SE_TM9SF3,SE_PIK3AP1
#> 6                                                                                           SE_CHD9
#>   DIST_TO_ANNOTATION_B FILTER
#> 1               -26937   PASS
#> 2                    0   PASS
#> 3                37570   PASS
#> 4               -22403   PASS
#> 5              -333830   PASS
#> 6                46382   PASS

# get all SVs for just one cohort
cohort_meta = suppressMessages(get_gambl_metadata()) %>% 
              dplyr::filter(cohort == "DLBCL_cell_lines")

some_sv <- get_combined_sv(these_samples_metadata = cohort_meta)
head(some_sv)
#> genomic_data Object
#> Genome Build: grch37 
#> Showing first 10 rows:
#>   CHROM_A   START_A     END_A CHROM_B  START_B    END_B
#> 1       1  23981141  23981142       1 60870954 60870955
#> 2       1  44208923  44208924       6 95573698 95573699
#> 3       1  62864466  62864736      15 27630123 27630439
#> 4       1  66804419  66804420      10 54466941 54466942
#> 5       1 105052013 105052280      20  9572155  9572453
#> 6       1 105428638 105428639       7 93357377 93357378
#>                   manta_name SCORE STRAND_A STRAND_B tumour_sample_id
#> 1 MantaBND:14444:0:1:0:0:0:0    NA        +        +           DOHH-2
#> 2 MantaBND:15937:0:1:0:0:0:1    NA        +        -           DOHH-2
#> 3 MantaBND:17095:0:1:0:0:0:0    NA        +        -           DOHH-2
#> 4 MantaBND:17346:0:2:0:0:0:0    NA        -        +           DOHH-2
#> 5   MantaBND:738:0:1:0:0:0:0    NA        -        +           DOHH-2
#> 6 MantaBND:20099:0:1:1:0:0:1    NA        +        -           DOHH-2
#>   normal_sample_id VAF_tumour DP gridss_name
#> 1        14-11247N      0.082 97        <NA>
#> 2        14-11247N      0.141 64        <NA>
#> 3        14-11247N      0.130 46        <NA>
#> 4        14-11247N      0.195 87        <NA>
#> 5        14-11247N      0.158 38        <NA>
#> 6        14-11247N      0.179 39        <NA>
#>                                                               ANNOTATION_A
#> 1 SE_AL021154.1,SE_AL450043.1,SE_E2F2,SE_ID3,SE_MDS2,SE_AL450043.1,SE_MDS2
#> 2                          SE_AC099795.1,SE_SLC2A1,SE_SLC2A1-AS1,SE_SLC2A1
#> 3                                                SE_JAK1,SE_RAVER2,SE_JAK1
#> 4                                                                 SE_PDE4B
#> 5                                                                 SE_NTNG1
#> 6                                                                 SE_NTNG1
#>   DIST_TO_ANNOTATION_A              ANNOTATION_B DIST_TO_ANNOTATION_B FILTER
#> 1                    0 SE_JAK1,SE_RAVER2,SE_JAK1              4422478   PASS
#> 2              -781402             SE_AL132996.1             -4465629   PASS
#> 3              2428697         SE_FAN1,SE_MTMR10              3583514   PASS
#> 4                    0                  SE_SGMS1             -2341951   PASS
#> 5              2845161                  SE_SHLD1             -3783469   PASS
#> 6              2468802             SE_AC003092.1               310993   PASS
nrow(some_sv)
#> [1] 3519

# get the SVs in a region around MYC
# WARNING: This is not the best way to find MYC SVs.
# Use annotate_sv on the full SV set instead.
myc_region_hg38 = "chr8:127710883-127761821"
myc_region_grch37 = "8:128723128-128774067"

hg38_myc_locus_sv <- get_combined_sv(region = myc_region_hg38,
                                projection = "hg38")
head(hg38_myc_locus_sv)
#> genomic_data Object
#> Genome Build: hg38 
#> Showing first 10 rows:
#>   CHROM_A   START_A     END_A CHROM_B   START_B     END_B
#> 1   chr14 106062935 106063377    chr8 127737727 127738231
#> 2   chr14 106063072 106063075    chr8 127737852 127737855
#> 3   chr14 105859407 105859410    chr8 127735901 127735904
#> 4   chr14 105859777 105859778    chr8 127735884 127735885
#> 5   chr14 105745498 105745499    chr8 127735666 127735667
#> 6   chr14 105857936 105857939    chr8 127737109 127737112
#>                    manta_name SCORE STRAND_A STRAND_B tumour_sample_id
#> 1 MantaBND:109733:0:1:1:0:0:1    NA        +        -       H059_T_WGS
#> 2 MantaBND:109733:0:1:0:2:0:1    NA        -        +       H059_T_WGS
#> 3 MantaBND:138918:0:1:0:0:0:1    NA        +        -          SP59270
#> 4 MantaBND:138918:0:1:1:0:0:1    NA        -        +          SP59270
#> 5 MantaBND:113362:0:1:0:0:0:1    NA        -        +          SP59272
#> 6 MantaBND:116726:1:2:0:0:0:1    NA        +        -          SP59276
#>   normal_sample_id VAF_tumour DP       gridss_name ANNOTATION_A
#> 1       H059_N_WGS      0.556  9  gridss152bf_237h      IGHV3-7
#> 2       H059_N_WGS      0.400 15 gridss152fb_1553h      IGHV3-7
#> 3          SP59269      0.350 20  gridss152bf_470h          Emu
#> 4          SP59269      0.786 14  gridss152fb_859h          Emu
#> 5          SP59271      0.293 41  gridss152fb_638h        IGHG1
#> 6          SP59274      0.500 24 gridss152bf_1277h          Emu
#>   DIST_TO_ANNOTATION_A ANNOTATION_B DIST_TO_ANNOTATION_B FILTER
#> 1                 -254          MYC                    0   PASS
#> 2                 -391          MYC                    0   PASS
#> 3                    0          MYC                  328   PASS
#> 4                    0          MYC                  347   PASS
#> 5                -2428          MYC                  565   PASS
#> 6                    0          MYC                    0   PASS
nrow(hg38_myc_locus_sv)
#> [1] 683

incorrect_myc_locus_sv <- get_combined_sv(region = myc_region_grch37,
                                projection = "hg38")
head(incorrect_myc_locus_sv)
#> genomic_data Object
#> Genome Build: hg38 
#> Showing first 10 rows:
#>   CHROM_A   START_A     END_A CHROM_B   START_B     END_B
#> 1    chr8 128738980 128738981    chr8 128752584 128752585
#> 2    chr8 127313080 127313571    chr8 128746007 128746587
#> 3    chr8 128738979 128738981    chr8 128752583 128752585
#> 4   chr11  93629111  93629647    chr8 128726343 128727380
#> 5   chr12 116681066 116681327    chr8 128759615 128759893
#> 6   chr20  36675075  36675080    chr8 128757284 128757289
#>                      manta_name  SCORE STRAND_A STRAND_B
#> 1 MantaDEL:330987:0:1:0:0:0_bp1     NA        +        -
#> 2 MantaDEL:204533:0:1:0:0:0_bp1     NA        +        -
#> 3 MantaDEL:136949:0:1:0:0:0_bp1     NA        +        -
#> 4    MantaBND:28037:1:9:0:0:0:1     NA        +        -
#> 5   MantaBND:295764:0:1:0:0:0:0     NA        -        -
#> 6                          <NA> 201.75        +        +
#>            tumour_sample_id          normal_sample_id VAF_tumour  DP
#> 1                SMZL-UK-T3                   SP59382 0.34600000  78
#> 2                  PD26401c                  PD26401b 0.07700000 259
#> 3                  SU-DHL-6                 14-11247N 0.46600000  58
#> 4                 01-20774T                 14-11247N 0.28000000  25
#> 5 HTMCP-01-01-00003-01D-03D HTMCP-01-01-00003-11A-01D 0.08700000 104
#> 6 HTMCP-01-16-00265-01A-01E HTMCP-01-16-00265-10A-01D 0.09195402  87
#>         gridss_name
#> 1              <NA>
#> 2              <NA>
#> 3              <NA>
#> 4              <NA>
#> 5              <NA>
#> 6 gridss152ff_7493h
#>                                                                                  ANNOTATION_A
#> 1                                                                                         MYC
#> 2                                                                                         MYC
#> 3                                                                                         MYC
#> 4                                                    SE_CEP295,SE_MIR1304,SE_SCARNA9,SE_TAF1D
#> 5                                                                      SE_AC083806.3,SE_RNFT2
#> 6 SE_DLGAP4-AS1,SE_RAB5IF,SE_SLA2,SE_TGIF2,SE_TGIF2-RAB5IF,SE_RAB5IF,SE_TGIF2,SE_TGIF2-RAB5IF
#>   DIST_TO_ANNOTATION_A ANNOTATION_B DIST_TO_ANNOTATION_B FILTER
#> 1              -996030          MYC             -1009634   PASS
#> 2               422662          MYC             -1003058   PASS
#> 3              -996030          MYC             -1009634   PASS
#> 4                34502          MYC              -983394   PASS
#> 5               117572          MYC             -1016666   PASS
#> 6               -62426          MYC             -1014335   PASS
nrow(incorrect_myc_locus_sv)
#> [1] 52

# Despite potentially being incomplete, we can nonetheless
# annotate these directly for more details
annotated_myc_hg38 = suppressMessages(
         GAMBLR.utils::annotate_sv(hg38_myc_locus_sv, genome_build = "hg38")
)
head(annotated_myc_hg38)
#>   chrom1    start1      end1 chrom2    start2      end2 name score strand1
#> 1      8 127712273 127712274      8 127766237 127766238    .    NA       +
#> 2      8 127226858 127226862      8 127759780 127759784    .    NA       +
#> 3      8 127226859 127226861      8 127759820 127759822    .    NA       -
#> 4      8 127736365 127736844      X  42943128  42943587    .    NA       +
#> 5      8 127736585 127736587      X  42943308  42943310    .    NA       -
#> 6     18  63195263  63195266      8 127744561 127744564    .    NA       -
#>   strand2 tumour_sample_id gene partner  fusion
#> 1       +  CAR_235_PreCART  MYC    <NA>  NA-MYC
#> 2       +          SP13307  MYC    <NA>  NA-MYC
#> 3       -          SP13307  MYC    <NA>  NA-MYC
#> 4       +         SP124957  MYC    <NA>  NA-MYC
#> 5       -         SP124957  MYC    <NA>  NA-MYC
#> 6       -         SP194216 BCL2    <NA> NA-BCL2
table(annotated_myc_hg38$partner)
#> 
#>  BCL6 CCNL1   DMD   IGH   IGK   IGL  LRMP  PAX5 RFTN1  ZEB2 
#>     6     1     2   377     5     7     3    11     1     1 
# The usual MYC partners are seen here

annotated_myc_incorrect = suppressMessages(
         GAMBLR.utils::annotate_sv(incorrect_myc_locus_sv, genome_build = "hg38")
)
head(annotated_myc_incorrect)
#>   chrom1    start1      end1 chrom2    start2      end2 name score strand1
#> 1      8 128738980 128738981      8 128752584 128752585    .    NA       +
#> 2      8 127313080 127313571      8 128746007 128746587    .    NA       +
#> 3      8 128738979 128738981      8 128752583 128752585    .    NA       +
#> 4      8 128757288 128757293      9  32897903  32897908    .    NA       +
#> 5      8 128738977 128738983      8 128752583 128752589    .    NA       +
#> 6      8 128738977 128738983      8 128752583 128752589    .    NA       +
#>   strand2 tumour_sample_id gene partner fusion
#> 1       -       SMZL-UK-T3  MYC    <NA> NA-MYC
#> 2       -         PD26401c  MYC    <NA> NA-MYC
#> 3       -         SU-DHL-6  MYC    <NA> NA-MYC
#> 4       -  01-23942_tumorB  MYC    <NA> NA-MYC
#> 5       -  04-14093_tumorA  MYC    <NA> NA-MYC
#> 6       -  04-14093_tumorB  MYC    <NA> NA-MYC
table(annotated_myc_incorrect$partner)
#> < table of extent 0 >
# The effect of specifying the wrong coordinate is evident