Get gene fusions

Retrieve fusions identified by FusionCatcher

Usage

get_gene_fusions(
  projection = "grch37",
  verbose = F,
  keep_genes,
  drop_genes,
  remove_adjacent_pairs = FALSE,
  these_samples_metadata,
  join_with_metadata = FALSE,
  harmonize_immunoglobulin_partners = FALSE
)

Arguments

projection

Which genome build

verbose

How chatty you want this experience

keep_genes

Vector of genes to restrict to

drop_genes

Vector of genes to drop (i.e. bad/artifacts)

remove_adjacent_pairs

Set to TRUE to drop genes that are already in close proximity

these_samples_metadata

The metadata to link to your fusion table

join_with_metadata

Set to TRUE to get fusions along with the sample metadata in one data frame

harmonize_immunoglobulin_partners

Attempt to clean up IG names so they are all consistently one of IGH, IGK or IGL

Value

A data frame in a bedpe-like format

Examples

all_fusions = get_gene_fusions()
onco_fusions = get_gene_fusions(keep_genes = c("BCL2","MYC","BCL6"))
print(head(onco_fusions))
#> # A tibble: 6 × 13
#>   sample_id CHROM_A   START_A     END_A CHROM_B START_B  END_B gene1 gene2 SCORE
#>   <chr>     <chr>       <dbl>     <dbl> <chr>     <dbl>  <dbl> <chr> <chr> <dbl>
#> 1 01-14875T 3       187463198 187463199 14       1.06e8 1.06e8 BCL6  IGH@      0
#> 2 01-15092T 18       60793477  60793478 14       1.06e8 1.06e8 BCL2  IGH@      0
#> 3 01-17838T 14      106070093 106070094 8        1.29e8 1.29e8 IGH@  MYC       0
#> 4 01-19969T 18       60794689  60794690 14       1.06e8 1.06e8 BCL2  IGH@      0
#> 5 01-20260T 18       60891187  60891188 14       1.06e8 1.06e8 BCL2  IGH@      0
#> 6 01-20260T 14      106209408 106209409 18       6.08e7 6.08e7 IGH@  BCL2      0
#> # ℹ 3 more variables: STRAND_A <chr>, STRAND_B <chr>, FLAGS <chr>