Get SSM By Sample.

Get the SSMs (i.e. load MAF) for a single sample.

get_ssm_by_sample(
  these_samples_metadata,
  tool_name = "slms-3",
  projection = "grch37",
  augmented = TRUE,
  flavour = "clustered",
  min_read_support = 3,
  basic_columns = TRUE,
  maf_cols = NULL,
  verbose = FALSE,
  this_sample_id,
  this_seq_type
)

Arguments

these_samples_metadata

Required if not specifying both this_sample_id and this_seq_type a single row or entire metadata table containing your sample_id.

tool_name

The name of the variant calling pipeline (currently only slms-3 is supported).

projection

The projection genome build. Supports hg38 and grch37.

augmented

default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead of the augmented MAF.

flavour

Currently this function only supports one flavour option but this feature is meant for eventual compatibility with additional variant calling parameters/versions.

min_read_support

Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs).

basic_columns

Return first 43 columns of MAF rather than full details. Default is TRUE.

maf_cols

if basic_columns is set to FALSE, the user can specify what columns to be returned within the MAF. This parameter can either be a vector of indexes (integer) or a vector of characters.

verbose

Enable for debugging/noisier output.

this_sample_id

Deprecated. Inferred from these_samples_metadata

this_seq_type

Deprecated. Inferred from these_samples_metadata

Value

data frame in MAF format.

Details

This was implemented to allow flexibility because there are some samples that we may want to use a different set of variants than those in the main GAMBL merge. The current use case is to allow a force_unmatched output to be used to replace the SSMs from the merge for samples with known contamination in the normal. This will also be useful to apply a blacklist to individual MAFs when coupled with annotate_ssm_blacklist. Is this function not what you are looking for? Try one of the following, similar, functions; get_coding_ssm, get_coding_ssm_status, get_ssm_by_patients, get_ssm_by_samples, get_ssm_by_region, get_ssm_by_regions

Examples

if (FALSE) { # \dontrun{
 this_sample_df = GAMBLR.results:::get_ssm_by_sample(
                          these_samples_metadata = get_gambl_metadata() %>%
                                dplyr::filter(sample_id == "HTMCP-01-06-00485-01A-01D",
                                     seq_type == "genome"),
                         projection = "grch37")

 capture_meta = get_gambl_metadata() %>% dplyr::filter(seq_type == "capture")

 ssm_sample = GAMBLR.results:::get_ssm_by_sample(this_sample_id = "CASA0002_2015-03-10",
                               projection = "grch37",
                               augmented = T,
                               these_samples_metadata = capture_meta)
} # }